Update Int16 to Int32, show country on SwiftUI

+++ Asclepione/Entities/CumulativeVaccinations.swift
@@ -11,12 +11,12 @@ import CoreData
 
 protocol CumulativeVaccinationsEntity {
     var areaName: String? { get set }
-    var cumulativeFirstDoses: Int16 { get set }
-    var cumulativeSecondDoses: Int16 { get set }
-    var cumulativeThirdDoses: Int16 { get set }
+    var cumulativeFirstDoses: Int32 { get set }
+    var cumulativeSecondDoses: Int32 { get set }
+    var cumulativeThirdDoses: Int32 { get set }
     var date: Date? { get set }
     var id: String? { get set }
-    var cumulativeVaccinations: Int16 { get set }
+    var cumulativeVaccinations: Int32 { get set }
 }
 
 public class CumulativeVaccinations: NSManagedObject, CumulativeVaccinationsEntity {}
@@ -28,12 +28,12 @@ extension CumulativeVaccinations {
     }
 
     @NSManaged public var areaName: String?
-    @NSManaged public var cumulativeFirstDoses: Int16
-    @NSManaged public var cumulativeSecondDoses: Int16
-    @NSManaged public var cumulativeThirdDoses: Int16
+    @NSManaged public var cumulativeFirstDoses: Int32
+    @NSManaged public var cumulativeSecondDoses: Int32
+    @NSManaged public var cumulativeThirdDoses: Int32
     @NSManaged public var date: Date?
     @NSManaged public var id: String?
-    @NSManaged public var cumulativeVaccinations: Int16
+    @NSManaged public var cumulativeVaccinations: Int32
     
     static var entityName: String { return "CumulativeVaccinations" }
